samples: bluetooth: peripheral_hids_mouse: force rpa for directed adv
This change adds an additional flag to the directed advertising
parameters. This flag instructs BLE stack to use RPA in initiator
address field of directed advertising packets even when privacy
feature is disabled. This is necessary to make directed advertising
packets visible to the central devices that are using Zephyr BLE stack.

Ref: NCSDK-4647

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>