tests: unity: Fix compilation of example_test
Example did not compile because of difference between main() signature.
In zephyr it is expected to be 'void main(void)' but unity generates
'int main(void)'. It could be solved by sending upstream patch to unity
but that is external dependancy.

Solution is using main_name option of unity to give main custom name and
implement actual main function (aligned with zephyr prototype) in test
file.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>