doc: fix issues with doxygen's @ref usage
Remove @ref usage where it is not being used to refer to a valid kind of
element. Most were converted to @em or @p (when referring to parameters
in the same description block). A few @ref were correctly used
(referring to group/section/etc) but refer to symbols in another repo
which is currently not supported.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>