Bluetooth: Fix nRF53 multi-image for BT_LL_NRFXLIB
The LLPM sample requires the BT_LL_NRFXLIB controller, but the
multi-image does not select that controller as default for the hci_rpmsg
project.

This commit adds an overlay that selects the BT_LL_NRFXLIB as default
for the child image if CONFIG_BT_LL_NRFXLIB_DEFAULT is set for the
parent.

Signed-off-by: Thomas Stenersen <thomas.stenersen@nordicsemi.no>