lib: at_cmd: match `OK` or `ERROR` only at the end of the response
Matching `OK` with strstr() won't work in case we are reading out
a certificate for example, where the occurence of `OK` can be quite
common; in that case the response would be incorrectly terminated
making it impossible to obtain the whole payload.

Only match `OK` or `ERROR` in the response if they are at the end.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>