get_pulls_in_range: rework for save/restore
Rework the internal representations so the relevant data which we need
to look up for each commit can be persisted in a sqlite database.

Save one using a new --sqlite-db command line argument, which provides
the database file. An in-memory database is used if this is not
provided just to shake out the code, but a warning is emitted in this
case. In my personal experience, it's very easy to get rate limited by
GitHub, and using a database to avoid having to collate results from
different runs manually is well worth it.

Use the standard library sqlite3 module to create an object oriented
wrapper around the database file. The wrapper is usable as a context
object, so the programming model is roughly this:

    with ResultsDatabase('my-sqlite.db') as db:
        for query in my_query_list:
            previous_result = db.get_result(query.sha)
            if previous_result:
                continue

            result = fetch_result_from_network(query)
            db.add_result(result)

The database results are committed regardless of whether the for loop
exits normally or via exception.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>