net: fota_download: Add error cause to library interface
Earlier there was no way to know whether the library returned an
error because the update could not be downloaded or because the
update was rejected. This commit adds a new error cause field to
the event which can be used to determine for example that the
modem rejected the update and there's no point in trying to
download the same update again.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>