Bluetooth: Mesh: Split CTL server set callbacks
Implements the pattern set by the xyL and HSL models for the CTL model,
in which the overarching collection model does not have a separate
callback structure, but relies on the underlying models. This changes
the CTL server's INIT macro into taking two callback structure
arguments, and the CTL callback structure is removed.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>