samples: debug: ppi_trace: Update sample configuration
Due to recent changes in clock control and boot up time, sample
was not behaving as described. Sample is tracing LF clock started
event. Tracing is setup in main so sample must be configured in a
way that LF clock is not started before entering main. Softdevice
LL is blocking until clock is ready so it cannot be used.
Additionally, by default zephyr is blocking until system clock is
stable. Reconfigured to no wait mode. Finally, it has been noticed
that initialization of CC310 driver on nrf52840 takes hundreds of
milliseconds. CC3xx driver is disabled now.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>