samples: nrf9160: agps: Don't request time from modem
In some networks time is not provided to the device, so the AT
command reading it out will fail.
The time is not used by this application or the libraries it uses,
so it can be disabled to avoid errors.

This should fix failures to request A-GPS data from nRF Cloud
in some networks.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>