net: lib: aws_iot: Add check for client subscriptions
This patch adds functionality that checks if topics
subscribed to by the AWS IoT backend are confirmed in
corresponding SUBACKS. When all topics are acknowledged
the backend propogates the AWS_IOT_READY event.
Previously the AWS IoT library would propogate
AWS_IOT_EVT_READY for each SUBACK received. The patch
also adds automatic shadow request upon a connection
establishment. This is added to this commit because
this functionality depends on proper SUBACK handling.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>