net: lib: azure_iot_hub: Fix parsing of direct method request ID
The current direct method request ID parsing uses atoi() and stores
the value as an int variabel.
However, the request IDs appear to be integers on hexadecimal form,
so the current parsing method will only work for IDs 1 to 9.
This patch changes the request ID type to a null-terminated string
instead of integer. The Azure IoT Hub doesn't explicitly
state in its docs what format the request ID is, so a string appears
to be safest choice.

The Azure IoT Hub and Azure FOTA samples are also updated accordingly.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>