lib: bsdlib: nrf91_sockets: Add mutex that protects getaddrinfo
This patch addresses an issue occurring when multiple DNS requests
are called from different contexts at the same time. By protecting
the getaddrinfo call with a mutex we can be sure to know that a thread
cannot preempt another thread that has just called a DNS request and
potentially receiving the result from the first request.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>