applications: serial_lte_modem BUG-FIX FOTA interrupted by AT command
When FOTA download is on-going, issue any AT command will interrupt it.
This is due to the Host and File buffers need to be kept during full
procedure of downloading (downloat_client/http.c), otherwise following
HTTP Request will contain tampered content in header and get 400 error.
Fix is to move the two buffers from function-wise to global data area.

Signed-off-by: Jun Qing Zou <jun.qing.zou@nordicsemi.no>