drivers: gps: Initialize driver on POST_KERNEL level
Change initialization level to POST_KERNEL to fix an issue with
invalid GNSS socket being used if gps_init() is called when
long-blocking initialization on APPLICATION level takes place.
This typically happens if an LTE connection is automatically
established in the initialization, which can take minutes.

The GNSS interface is usable after BSDlib is initialized, and
this happens at POST_KERNEL level, with priority 0, while the GPS
driver is now set to have priority 90 by default.

Also initialize the socket to an invalid value for the same reasons.

Fixes CIA-135

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>