Bluetooth: Mesh: Disable light control on all lightness changes
Adds clearing of the CONTROLLED flag in the lightness server for all
lightness changing messages, both from the server's own messages and the
extended models' messages. This ensures compatibility with the Bluetooth
Mesh Model specification section 6.2.3.1, which states that all manual
changes to the lightness should disable the light control server.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>