Bluetooth: Mesh: Fix Light Controller cooldown timer
The manual mode cooldown duration calculation did not treat the manual
cooldown configuration as seconds, but milliseconds. It also failed to
account for underflow when subtracting the fade duration, and always
assumed that the fade happened with the configured standby manual
duration, which would not be the case if the transition was set
manually.

Reworks the timeout function in Light Controller to account for these
errors and reduce the amount of indentation.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>