cmake: correct signature key file handling for multi image builds
Fixes: NCSDK-7277

This commit ensure that the path to signature key files is properly
handled when an absolute path is provided.

It aligns handling of relative path to MCUBoot upstream when multiple
mcuboot_CONF_FILE are provided as list.

The PEM may contain just a public key to be embedded into mcuboot,
therefore a KConfig option to disable signing has been introduced.
Also it is checked that the key file contains a private key, and if not
then a warning is printed to the user, and no signing will be performed.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>